@@ -3,7 +3,7 @@ description: A Flutter application showcasing how to use the dynamic_fonts packa
 publish_to: none
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
