@@ -1,25 +1,23 @@
 name: dynamic_fonts
 description: A package to include dynamic web-hosted fonts in your Flutter app.
-version: 1.1.2
+version: 2.0.0-nullsafety.0
 homepage: https://github.com/amake/dynamic-fonts-flutter/
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
-  flutter: ^1.17.0
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.0+2
-  path_provider: '>=1.4.0 <3.0.0'
-  crypto: '>=2.1.3 <4.0.0'
+  http: ^0.13.0
+  path_provider: ^2.0.0-nullsafety
+  crypto: ^3.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.1
+  mockito: ^5.0.0-nullsafety
   mustache: ^1.1.1
   pedantic: ^1.8.0
   protobuf: ^1.0.1
-  protoc_plugin: any
   console: ^3.1.0
