@@ -2,17 +2,14 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter/foundation.dart';
-
 import 'google_fonts_variant.dart';
 
 /// Represents a Google Fonts API variant in Flutter-specific types.
 class GoogleFontsFamilyWithVariant {
   const GoogleFontsFamilyWithVariant({
-    @required this.family,
-    @required this.googleFontsVariant,
-  })  : assert(family != null),
-        assert(googleFontsVariant != null);
+    required this.family,
+    required this.googleFontsVariant,
+  });
 
   final String family;
   final GoogleFontsVariant googleFontsVariant;
