@@ -12,11 +12,11 @@ import 'package:dynamic_fonts/src/asset_manifest.dart';
 const _fakeAssetManifestText = '{"value": ["fake"]}';
 var _assetManifestLoadCount = 0;
 
-AssetManifest assetManifest;
+late AssetManifest assetManifest;
 
 void main() {
   setUpAll(() async {
-    ServicesBinding.instance.defaultBinaryMessenger
+    ServicesBinding.instance!.defaultBinaryMessenger
         .setMockMessageHandler('flutter/assets', (message) {
       _assetManifestLoadCount++;
       final Uint8List encoded = utf8.encoder.convert(_fakeAssetManifestText);
@@ -34,10 +34,10 @@ void main() {
 
   testWidgets('AssetManifest loads once when called multiple times in parallel',
       (tester) async {
-    final manifestJsons = await Future.wait([
-      assetManifest.json(),
-      assetManifest.json(),
-      assetManifest.json(),
+    final manifestJsons = await Future.wait<Map<String, List<String>>?>([
+      assetManifest.json()!,
+      assetManifest.json()!,
+      assetManifest.json()!,
     ]);
     _verifyAssetManifestLoadedOnce();
     manifestJsons.forEach(_verifyAssetManifestContent);
@@ -46,10 +46,10 @@ void main() {
   testWidgets(
       'AssetManifest loads once when called multiple times in parallel then multiple times in succession',
       (tester) async {
-    final manifestJsons = await Future.wait([
-      assetManifest.json(),
-      assetManifest.json(),
-      assetManifest.json(),
+    final manifestJsons = await Future.wait<Map<String, List<String>>?>([
+      assetManifest.json()!,
+      assetManifest.json()!,
+      assetManifest.json()!,
     ]);
     _verifyAssetManifestLoadedOnce();
     manifestJsons.forEach(_verifyAssetManifestContent);
@@ -84,6 +84,6 @@ void _verifyAssetManifestLoadedOnce() {
   expect(_assetManifestLoadCount, 1);
 }
 
-void _verifyAssetManifestContent(Map<String, dynamic> manifestJson) {
-  expect(manifestJson['value'], ['fake']);
+void _verifyAssetManifestContent(Map<String, dynamic>? manifestJson) {
+  expect(manifestJson!['value'], ['fake']);
 }
