@@ -14,7 +14,12 @@ import 'package:http/http.dart' as http;
 import 'package:mockito/mockito.dart';
 import 'package:path_provider/path_provider.dart';
 
-class MockHttpClient extends Mock implements http.Client {}
+class MockHttpClient extends Mock implements http.Client {
+  Future<http.Response> gets(dynamic uri, {dynamic headers}) {
+    super.noSuchMethod(Invocation.method(#get, [uri], {#headers: headers}));
+    return Future.value(http.Response('', 200));
+  }
+}
 
 class MockAssetManifest extends Mock implements AssetManifest {}
 
@@ -40,13 +45,16 @@ void overridePrint(Future<Null> testFn()) => () {
     };
 
 void main() {
-  Directory directory;
+  late Directory directory;
+
+  late MockHttpClient _httpClient;
 
   setUp(() async {
-    httpClient = MockHttpClient();
+    _httpClient = MockHttpClient();
+    httpClient = _httpClient;
     assetManifest = MockAssetManifest();
     DynamicFonts.config.allowRuntimeFetching = true;
-    when(httpClient.get(any)).thenAnswer((_) async {
+    when(_httpClient.gets(any)).thenAnswer((_) async {
       return http.Response(_fakeResponse, 200);
     });
 
@@ -81,13 +89,13 @@ void main() {
 
     await loadFontIfNecessary(fakeDescriptor);
 
-    verify(httpClient.get(anything)).called(1);
+    verify(_httpClient.gets(anything)).called(1);
   });
 
   testWidgets('loadFontIfNecessary method throws if font cannot be loaded',
       (tester) async {
     // Mock a bad response.
-    when(httpClient.get(any)).thenAnswer((_) async {
+    when(_httpClient.gets(any)).thenAnswer((_) async {
       return http.Response('fake response body - failure', 300);
     });
 
@@ -143,7 +151,7 @@ void main() {
       );
     });
 
-    verifyNever(httpClient.get(anything));
+    verifyNever(_httpClient.gets(anything));
   });
 
   testWidgets(
@@ -162,15 +170,15 @@ void main() {
 
     // 1st call.
     await loadFontIfNecessary(fakeDescriptor);
-    verify(httpClient.get(anything)).called(1);
+    verify(_httpClient.gets(anything)).called(1);
 
     // 2nd call.
     await loadFontIfNecessary(fakeDescriptor);
-    verifyNever(httpClient.get(anything));
+    verifyNever(_httpClient.gets(anything));
 
     // 3rd call.
     await loadFontIfNecessary(fakeDescriptor);
-    verifyNever(httpClient.get(anything));
+    verifyNever(_httpClient.gets(anything));
   });
 
   testWidgets(
@@ -192,7 +200,7 @@ void main() {
       loadFontIfNecessary(fakeDescriptor),
       loadFontIfNecessary(fakeDescriptor)
     ]);
-    verify(httpClient.get(anything)).called(1);
+    verify(_httpClient.gets(anything)).called(1);
   });
 
   testWidgets(
@@ -210,16 +218,16 @@ void main() {
     );
 
     // Have the first call throw an error.
-    when(httpClient.get(any)).thenThrow('error');
+    when(_httpClient.gets(any)).thenThrow('error');
     await loadFontIfNecessary(fakeDescriptor);
-    verify(httpClient.get(any)).called(1);
+    verify(_httpClient.gets(any)).called(1);
 
     // The second call will retry the http fetch.
-    when(httpClient.get(any)).thenAnswer((_) async {
+    when(_httpClient.gets(any)).thenAnswer((_) async {
       return http.Response(_fakeResponse, 200);
     });
     await loadFontIfNecessary(fakeDescriptor);
-    verify(httpClient.get(any)).called(1);
+    verify(_httpClient.gets(any)).called(1);
   });
 
   testWidgets('loadFontIfNecessary method writes font file', (tester) async {
@@ -234,12 +242,12 @@ void main() {
     );
 
     var directoryContents = await getApplicationSupportDirectory();
-    expect(directoryContents.listSync().isEmpty, isTrue);
+    expect(directoryContents!.listSync().isEmpty, isTrue);
 
     await loadFontIfNecessary(fakeDescriptor);
     directoryContents = await getApplicationSupportDirectory();
 
-    expect(directoryContents.listSync().isNotEmpty, isTrue);
+    expect(directoryContents!.listSync().isNotEmpty, isTrue);
     expect(
       directoryContents.listSync().single.toString().contains('Foo'),
       isTrue,
@@ -249,7 +257,7 @@ void main() {
   testWidgets(
       'loadFontIfNecessary does not save anything to disk if the file does not '
       'match the expected hash', (tester) async {
-    when(httpClient.get(any)).thenAnswer((_) async {
+    when(_httpClient.gets(any)).thenAnswer((_) async {
       return http.Response('malicious intercepted response', 200);
     });
     final fakeDescriptor = GoogleFontsDescriptor(
@@ -264,10 +272,10 @@ void main() {
     );
 
     var directoryContents = await getApplicationSupportDirectory();
-    expect(directoryContents.listSync().isEmpty, isTrue);
+    expect(directoryContents!.listSync().isEmpty, isTrue);
 
     await loadFontIfNecessary(fakeDescriptor);
     directoryContents = await getApplicationSupportDirectory();
-    expect(directoryContents.listSync().isEmpty, isTrue);
+    expect(directoryContents!.listSync().isEmpty, isTrue);
   });
 }
