@@ -1,20 +1,20 @@
 import 'dart:io';
 import 'dart:typed_data';
 
-import 'package:path_provider/path_provider.dart' as path_provider;
+import 'package:path_provider/path_provider.dart';
 
 Future<void> saveFontToDeviceFileSystem(String name, List<int> bytes) async {
   final file = await _localFile(name);
   await file.writeAsBytes(bytes);
 }
 
-Future<ByteData> loadFontFromDeviceFileSystem(String name) async {
+Future<ByteData?> loadFontFromDeviceFileSystem(String name) async {
   try {
     final file = await _localFile(name);
     final fileExists = file.existsSync();
     if (fileExists) {
       List<int> contents = await file.readAsBytes();
-      if (contents != null && contents.isNotEmpty) {
+      if (contents.isNotEmpty) {
         return ByteData.view(Uint8List.fromList(contents).buffer);
       }
     }
@@ -25,8 +25,8 @@ Future<ByteData> loadFontFromDeviceFileSystem(String name) async {
 }
 
 Future<String> get _localPath async {
-  final directory = await path_provider.getApplicationSupportDirectory();
-  return directory.path;
+  final directory = await getApplicationSupportDirectory();
+  return directory!.path;
 }
 
 Future<File> _localFile(String name) async {
